New Features

+++ src/public-toilet/public-toilet.service.ts
@@ -1,5 +1,9 @@
-import { GeoJSONPointScalar } from "src/common/scalars/geojson/Point.scalar";
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { GeoJSONPointScalar } from 'src/common/scalars/geojson/Point.scalar';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreatePublicToiletInput } from './dto/create-public-toilet.input';
@@ -9,6 +13,10 @@ import {
   PublicToiletImage,
 } from './entities/public-toilet.entity';
 import { uploadFileStream } from '../common/utils/upload';
+import { PublishState as PublicToiletState } from 'src/common/enum/publish_state.enum';
+import { checkUserIsAuthor } from 'src/common/utils/checkUserAuthentication';
+import { checkIfObjectIsPublished } from 'src/common/utils/checkPublishedState';
+import { FilesService } from 'src/common/services/files.service';
 
 @Injectable()
 export class PublicToiletService {
@@ -17,12 +25,13 @@ export class PublicToiletService {
     private publicToiletRepository: Repository<PublicToilet>,
     @InjectRepository(PublicToiletImage)
     private publicToiletImageRepository: Repository<PublicToiletImage>,
+    private fileService: FilesService,
   ) {}
   uploadDir = process.env.MEDIA_ROOT;
 
   async create(
     publicToiletInput: CreatePublicToiletInput,
-    user: number
+    user: number,
   ): Promise<PublicToilet> {
     let publicToiletInputData = {
       ...publicToiletInput,
@@ -46,7 +55,10 @@ export class PublicToiletService {
     const publicToiletData: PublicToilet =
       await this.publicToiletRepository.save({
         ...publicToiletInputData,
-        publishedAt: new Date(),
+        publishedAt:
+          publicToiletInput.state === PublicToiletState.PUBLISHED
+            ? new Date()
+            : null,
         createdAt: new Date(),
         updatedAt: new Date(),
         createdBy: user,
@@ -82,16 +94,48 @@ export class PublicToiletService {
       };
     }
 
-    return await publicToiletData;
+    return publicToiletData;
   }
 
   async findAll(
     limit: number,
     offset: number,
+    publishedOnly = false,
   ): Promise<[PublicToilet[], number]> {
+    const whereOptions: any = {};
+    if (publishedOnly) {
+      whereOptions.state = PublicToiletState.PUBLISHED;
+    }
     return this.publicToiletRepository.findAndCount({
       take: limit,
       skip: offset,
+      where: {
+        ...whereOptions,
+      },
+      order: {
+        createdAt: 'DESC',
+      },
+    });
+  }
+
+  async findAllAdmin(
+    limit: number,
+    offset: number,
+    publishedOnly = false,
+  ): Promise<[PublicToilet[], number]> {
+    const whereOptions: any = {};
+    if (publishedOnly) {
+      whereOptions.state = PublicToiletState.PUBLISHED;
+    }
+    return this.publicToiletRepository.findAndCount({
+      take: limit,
+      skip: offset,
+      where: {
+        ...whereOptions,
+      },
+      order: {
+        createdAt: 'DESC',
+      },
     });
   }
 
@@ -106,7 +150,11 @@ export class PublicToiletService {
     return new NotFoundException(`PublicToilet with id ${id} not found`);
   }
 
-  async update(id: number, updatePublicToiletInput: UpdatePublicToiletInput, user: number) {
+  async update(
+    id: number,
+    updatePublicToiletInput: UpdatePublicToiletInput,
+    user: number,
+  ) {
     const publicToilet: PublicToilet =
       await this.publicToiletRepository.findOne({
         where: { id: id },
@@ -160,6 +208,11 @@ export class PublicToiletService {
       const { images, ...updatedPublicToilet } = {
         ...publicToilet,
         ...publicToiletInputData,
+        publishedAt:
+          publicToilet.publishedAt ||
+          (updatePublicToiletInput.state === PublicToiletState.PUBLISHED
+            ? new Date()
+            : null),
         updatedAt: new Date(),
         updatedBy: user,
       };
@@ -168,7 +221,7 @@ export class PublicToiletService {
     throw new NotFoundException(`PublicToilet with id ${id} not found`);
   }
 
-  async remove(id: number) {
+  async remove(id: number, user: number) {
     const publicToilet: PublicToilet =
       await this.publicToiletRepository.findOne({
         where: { id: id },
@@ -176,6 +229,13 @@ export class PublicToiletService {
       });
 
     if (publicToilet) {
+      const published_state = checkIfObjectIsPublished(publicToilet.state);
+      if (published_state) {
+        return new ForbiddenException(
+          `PublicToilet with id ${id} already published.`,
+        );
+      }
+      checkUserIsAuthor(user, publicToilet.createdBy);
       const deleteImage = publicToilet.images.map(async (image) => {
         return await this.publicToiletImageRepository.delete(image.id);
       });
@@ -184,6 +244,19 @@ export class PublicToiletService {
       return publicToilet;
     }
     return new NotFoundException(`PublicToilet with id ${id} not found`);
+  }
+
+  async removeImage(id: number) {
+    const image = await this.publicToiletImageRepository.findOne({
+      where: { id: id },
+    });
+    if (image) {
+      await this.publicToiletImageRepository.delete(image.id);
+      this.fileService.removeFile(image.image);
+      return image;
+    }
+
+    return new NotFoundException(`Image with id ${id} not found`);
   }
 
   async findImagesOfPublicToilet(id: number) {
@@ -198,19 +271,25 @@ export class PublicToiletService {
     return new NotFoundException(`PublicToilet with id ${id} not found`);
   }
 
-  async findPublicToiletNearMe(limit:number, offset:number, origin: GeoJSONPointScalar): Promise<[PublicToilet[], number]> {
-      let distance = 40;
-      const publicToilet = await this.publicToiletRepository
+  async findPublicToiletNearMe(
+    limit: number,
+    offset: number,
+    origin: GeoJSONPointScalar,
+  ): Promise<[PublicToilet[], number]> {
+    let distance = 40;
+    const publicToilet = await this.publicToiletRepository
       .createQueryBuilder()
-      .where('ST_DWithin(geopoint, ST_SetSRID(ST_GeomFromGeoJSON(:origin), ST_SRID(geopoint)), :distance)')
+      .where('state = :state', { state: PublicToiletState.PUBLISHED })
+      .andWhere(
+        'ST_DWithin(geopoint, ST_SetSRID(ST_GeomFromGeoJSON(:origin), ST_SRID(geopoint)), :distance)',
+      )
       .setParameters({
-          origin: JSON.stringify(origin),
-          distance: distance*1000, //KM conversion
+        origin: JSON.stringify(origin),
+        distance: distance * 1000, //KM conversion
       })
       .limit(limit)
       .offset(offset)
       .getMany();
-      return [publicToilet, publicToilet.length];
+    return [publicToilet, publicToilet.length];
   }
 }
-
